requirements.txt
pillow>=10.0.0
pydantic>=2.0.0
anthropic>=0.40.0
websocket-client>=1.6.0
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
structlog>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
